[feat] #39 preview 추가 , package 이동
