+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/onboarding/navigation/OnboardingNavigation.kt
@@ -9,10 +9,10 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import com.dpm.sixpack.presentation.destinations.OnboardingRoute
 import com.dpm.sixpack.presentation.routes.onboarding.OnboardingViewModel
-import com.dpm.sixpack.presentation.routes.onboarding.routes.finish.OnboardingFinishRoute
-import com.dpm.sixpack.presentation.routes.onboarding.routes.goal.OnboardingGoalRoute
-import com.dpm.sixpack.presentation.routes.onboarding.routes.level.OnboardingLevelRoute
-import com.dpm.sixpack.presentation.routes.onboarding.routes.permission.OnboardingPermissionRoute
+import com.dpm.sixpack.presentation.routes.onboarding.OnboardingFinishRoute
+import com.dpm.sixpack.presentation.routes.onboarding.OnboardingGoalRoute
+import com.dpm.sixpack.presentation.routes.onboarding.OnboardingLevelRoute
+import com.dpm.sixpack.presentation.routes.onboarding.OnboardingPermissionRoute
 
 fun NavController.navigateOnboarding(navOptions: NavOptions? = null) {
     navigate(OnboardingRoute.Onboarding, navOptions)
