+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/onboarding/ui/screen/finish/OnboardingFinishScreen.kt
@@ -1,6 +1,5 @@
-package com.dpm.sixpack.presentation.routes.onboarding.routes.finish
+package com.dpm.sixpack.presentation.routes.onboarding.ui.screen.finish
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -8,24 +7,21 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.dpm.sixpack.presentation.R
 import com.dpm.sixpack.presentation.common.components.goal.RecommendedGoalList
 import com.dpm.sixpack.presentation.common.components.preview.DoRunPreviewWrapper
 import com.dpm.sixpack.presentation.common.components.topbar.DoRunNavigationTopBar
-import com.dpm.sixpack.presentation.routes.onboarding.component.OnboardingNextButton
+import com.dpm.sixpack.presentation.routes.onboarding.ui.component.common.OnboardingNextButton
 import com.dpm.sixpack.presentation.routes.onboarding.contract.uistate.OnboardingUiState
 import com.dpm.sixpack.presentation.common.components.goal.model.state.GoalUiState
 import com.dpm.sixpack.presentation.common.components.goal.model.state.RecommendedGoalUiState
 import com.dpm.sixpack.presentation.common.components.goal.model.type.GoalType
+import com.dpm.sixpack.presentation.routes.onboarding.ui.component.finish.FinishScreenHeader
 import com.dpm.sixpack.presentation.theme.SixpackTheme
 
 @Composable
@@ -76,33 +72,6 @@ fun OnboardingFinishScreen(
                 modifier = Modifier.padding(horizontal = 20.dp),
             )
         }
-    }
-}
-
-@Composable
-private fun FinishScreenHeader(modifier: Modifier = Modifier) {
-    Column(
-        modifier =
-            modifier
-                .fillMaxWidth(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Image(
-            imageVector = ImageVector.vectorResource(id = R.drawable.ill_goal_character),
-            contentDescription = null,
-            modifier = Modifier,
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = stringResource(R.string.onboarding_finish_title),
-            style = SixpackTheme.typography.h2Bold,
-            modifier =
-                Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 20.dp),
-        )
     }
 }
